\name{example_output_DMRforPairs}
\docType{data}
\alias{example_output_DMRforPairs}
\title{EXAMPLE OF DMRforPairs RESULTS}
\description{
  This dataset includes the DMRforPairs results of a 4MB region in the CL.methy dataset. This is a small subset of the data used in the vignette to facilitate the examples in the reference manual(104MB-107MB on chr 7). Please see the CL.methy dataset for more details.
}
\usage{example_output_DMRforPairs}
\format{DMRforPairs results formatted as described in the reference manual for the \code{\link{DMRforPairs}} wrapper.}
\source{Bibikova et al 2009 & 2011, \href{http://support.illumina.com/downloads/genomestudio_software_20111.ilmn}{download from Illumina website}.}
\keyword{datasets}