\name{plot_annotate_region}
\alias{plot_annotate_region}
\title{PLOT METHYLATION AND ANNOTATE REGION}
\description{
Wrapper for \code{\link{plot_annotate_probes}}. Generates a plot of the methylation status of a region identified by DMRforPairs and associated transcripts (if annotate=TRUE).
}
\usage{
plot_annotate_region(tested,regions,margin=10000,regionID,clr=NA,annotate=TRUE,scores=TRUE, path)
}
\arguments{
  \item{tested}{Output of \code{\link{testregion}}}
  \item{regions}{Output of \code{\link{regionfinder}}}
  \item{margin}{See \code{\link{export_data}}}
  \item{regionID}{ID of a region identified by DMRforPairs as presented in the output of \code{\link{testregion}} and the output generated by \code{\link{export_data}}.}
  \item{clr}{See \code{\link{export_data}}}
  \item{annotate}{If FALSE, only basic visualizations without annotation information are generated. (default=TRUE)}
  \item{scores}{Specifies if statistics need to be calculated for the region or just figures. In a custom request from a user, statistics are relevant to calculate. In the \code{\link{export_data}} wrapper, they are redundant (already calculated in the \code{\link{testregion}} loop). (default=TRUE)}
  \item{path}{Path relative to the working directory to save the files (=experiment.name in \code{\link{export_data}})}
}
\details{
Expects a subfolder with the name of your experiment to exist as created by export_data(). Writing permissions are required in the working directory to use the export and visualization functions of DMRforPairs. Internet access is required to use the annotation features.\cr
}
\value{
A number of figures  (including pairwise plots if n>2) is generated visualizing methylation status in the samples at / near the region investigated. Statistics and optional annotation information are reported back by the function.
}
\seealso{
\code{\link{export_data}}, 
\code{\link{plot_annotate_probes}}, 
\code{\link{regionfinder}}, 
\code{\link{testregion}}.
}
\author{Martin Rijlaarsdam, \email{m.a.rijlaarsdam@gmail.com}}
\examples{
#For the this step, please ensure that you have writing 
#permissions in your working directory. If any of the
#annotate parameters is set to TRUE, an internet connection
#is required.

#this function expects a folder "example_DMRforPairs" to exist and
#be writable in the current working directory.
data(DMRforPairs_data)
experiment="example_DMRforPairs"

plot_annotate_region(example_output_DMRforPairs$tested,
                     example_output_DMRforPairs$regions,
                     margin=10000,
                     regionID=7,
                     annotate=FALSE,scores=TRUE,path=experiment)   
}
\keyword{export_visualization}
